echo_prepare_agent/pyproject.toml
[project]
name = "echo-prepare-agent"
version = "0.1.0"
description = "Echo Prepare Agent - Student study and exam preparation assistant using Strands SDK"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk>=0.3.12",
    "bedrock-agentcore>=0.1.7",
    "strands-agents>=1.16.0",
    "strands-tools>=1.0.0",
    "tavily-python>=0.7.12",
    "uvicorn>=0.37.0",
    "python-dotenv>=1.0.0",
]
